accessible/tests/mochitest/test_childAtPoint.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="nsIAccessible::getChildAtPoint and getDeepestChildAtPoint">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

  <script type="application/javascript"
          src="treeview.js" />

  <script type="application/javascript"
          src="common.js" />
  <script type="application/javascript"
          src="layout.js" />

  <script type="application/javascript">
  <![CDATA[
    function doTest()
    {
      // Initialize the tree
      var view = new nsTableTreeView(5);

      var tree = getNode("tree");
      var treeBox = tree.treeBoxObject;
      treeBox.view = view;

      // Tests
      var treecols = getNode("treecols");
      var x = treecols.boxObject.x;
      var y = treecols.boxObject.y;

      testChildAtPoint(tree, x, y, false, treecols);
      testChildAtPoint(tree, x, y, true, "col1");

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  ]]>
  </script>

  <hbox flex="1" style="overflow: auto;">
    <body xmlns="http://www.w3.org/1999/xhtml">
      <a target="_blank"
         href="https://bugzilla.mozilla.org/show_bug.cgi?id=471493"
         title=" crash [@ nsPropertyTable::GetPropertyInternal]">
        Mozilla Bug 471493
      </a><br/>
      <p id="display"></p>
      <div id="content" style="display: none">
      </div>
      <pre id="test">
      </pre>
    </body>

    <tree id="tree" flex="1">
      <treecols id="treecols">
        <treecol id="col1" flex="1" primary="true" label="column"/>
        <treecol id="col2" flex="1" label="column 2"/>
      </treecols>
      <treechildren id="treechildren"/>
    </tree>
  </hbox>

</window>
